LOW: clean up
Small toml files and python codes clean up to prepare for adding the code to Pypi. README file was also updated.

+++ python/src/rsliding/utils.py
@@ -76,5 +76,5 @@ class BaseCheck:
         intended).
         """
 
-        self._data = np.ascontiguousarray(self._data)
-        self._kernel = np.ascontiguousarray(self._kernel)
+        self._data = np.ascontiguousarray(self._data, dtype=np.float64)
+        self._kernel = np.ascontiguousarray(self._kernel, dtype=np.float64)
